==> 1a_QC_doublets_integration_global.Rmd <==
---
title: "1a_QC_doublets_integration_global"
---

```{r}
getwd()
```

#1_lib

```{r}
library(Seurat)
library(scDblFinder)
library(tidyverse)
```

#2_RDS

```{r}
seurat <- readRDS("..... .rds")
```

#3_Doublets and mito

```{r}
sce <- as.SingleCellExperiment(seurat)
```

```{r}
sce <- scDblFinder(sce, samples="orig.ident")
```

```{r}
hist(sce$scDblFinder.score, breaks=100)
```

```{r}
seurat$Doublet <- sce$scDblFinder.score > 0.15 
```

```{r}
doubletfig <- DimPlot(seurat, group.by="Doublet")
doubletfig
#ggsave(doubletfig, filename = "... .png", width=12, height=8, dpi = 900)
```

```{r}
cells_keep <- colnames(seurat)[!seurat$Doublet]
```

```{r}
seurat <- subset(seurat, cells=cells_keep)
```

```{r}
FeaturePlot(seurat, features="percent.mito")
hist(seurat$percent.mito, breaks=100)
```

```{r}
seurat <- subset(seurat, subset= percent.mito < 20)
```

#4_Add Cell Cycle
```{r}
seurat <- CellCycleScoring(seurat, s.features=cc.genes.updated.2019$s.genes,
                           g2m.features = cc.genes.updated.2019$g2m.genes)
```

```{r}
cyclefig <- DimPlot(seurat, group.by="Phase")
ggsave(cyclefig, filename = ".../ CellCycle.png", width=12, height=8, dpi = 900)
cyclefig
```

#5_Integrate

```{r}
options(future.globals.maxSize = 4000 * 1024^2)
seurat.list <- SplitObject(seurat, split.by="Sample")
for (i in 1:length(seurat.list)){
  seurat.list[[i]] <- SCTransform(seurat.list[[i]], 
                                  vars.to.regress=c("percent.mito", "S.Score", "G2M.Score"),
                                  verbose=F)
}
seurat.features <- SelectIntegrationFeatures(object.list = seurat.list, nfeatures=3000)
seurat.list <- PrepSCTIntegration(object.list=seurat.list, 
                                  anchor.features=seurat.features, 
                                  verbose=F)
seurat.anchors <- FindIntegrationAnchors(object.list = seurat.list, normalization.method="SCT",
                                         anchor.features=seurat.features, verbose=F)
seurat.integrated <- IntegrateData(anchorset = seurat.anchors, normalization.method="SCT",
                                   verbose=F)

seurat <- seurat.integrated
rm(seurat.integrated)
seurat <- RunPCA(seurat, verbose=F)
seurat <- RunUMAP(seurat, dims=1:30)
seurat <- FindNeighbors(seurat, dims=1:30)
seurat <- FindClusters(seurat, resolution=0.25)
```

```{r}
DimPlot(seurat, label=T)
DimPlot(seurat, group.by="Sample")
```

#6_Re-embed in non-integrated space
```{r}
DefaultAssay(seurat) <- "RNA"
seurat <- SCTransform(seurat, vars.to.regress=c("percent.mito",
                                                "S.Score", "G2M.Score"),
                      verbose=F)
seurat <- RunPCA(seurat, verbose=F)
seurat <- RunUMAP(seurat, dims=1:30)
```

```{r}
DimPlot(seurat, label=T)
DimPlot(seurat, group.by="Sample")
```

```{r}
DefaultAssay(seurat) <- "RNA" # We use the RNA for all quantitative downstream analyses
```

```{r}
#saveRDS(seurat, file = "... .rds")
```
