==> 1c_mouse_global_single_cell_net.Rmd <==
---
title: "1c_mouse_global_single_cell_net"
---

#1_lib

```{r}
library(singleCellNet)
library(Seurat)
```

#2_Load Query data 

```{r}
seurat <- readRDS("... .rds")
```

```{r
seurat <- extractSeurat(seurat, exp_slot_name = "counts")
stPark = seurat$sampTab      
expPark = seurat$expDat      

genesPark =  rownames(expPark)

rm(expPark)

gc()        
```

#3_Load the training data 

This is the tabula muris complete data set 

```{r}
#loading training data
tm_10x <- utils_loadObject("/... .rda")

expTMraw= tm_10x$expDat
dim(expTMraw)

stTM = tm_10x$sampTab
dim(stTM)

stTM <- droplevels(stTM)
```

```{r}
table(stTM[which(stTM$tissue == "Lung"), "cell_ontology_class"])
```

```{r}
#Keep only lung cells
stTM = stTM[which(stTM$tissue == "Lung"), ]
dim(stTM)
```

```{r}
#eliminate NAs
stTM = stTM[-which(is.na(stTM$cell_ontology_class)),] #remove NAs
dim(stTM)
```

```{r}
commonGenes = intersect(rownames(expTMraw), genesPark)
length(commonGenes)
```

```{r}
expTMraw = expTMraw[commonGenes,]
```

#5_Split for training and assessment, and transform training data

```{r}
set.seed(100) #can be any random seed number
stList = splitCommon(sampTab=stTM, ncells=24, dLevel="cell_ontology_class")
stTrain = stList[[1]]
expTrain = expTMraw[,stTrain$cell] 
```

```{r}
system.time(class_info<-scn_train(stTrain = stTrain, expTrain = expTrain, nTopGenes = 10, nRand = 100, nTrees = 1000, nTopGenePairs = 25, dLevel = "cell_ontology_class", colName_samp = "cell"))
```

```{r}
#validate data
stTestList = splitCommon(sampTab=stList[[2]], ncells=100, dLevel="cell_ontology_class") 
stTest = stTestList[[1]]
expTest = as.matrix(expTMraw[,as.vector(stTest$cell)])

classRes_val_all = scn_predict(cnProc=class_info[['cnProc']], expDat=expTest, nrand = 50)

```
```{r}
tm_heldoutassessment = assess_comm(ct_scores = classRes_val_all, stTrain = stTrain, 
                                   stQuery = stTest, dLevelSID = "cell", 
                                   classTrain = "cell_ontology_class", classQuery = "cell_ontology_class", nRand = 50)

```

```{r}
plot_PRs(tm_heldoutassessment)
```

```{r}
plot_metrics(tm_heldoutassessment)
```

```{r}
#Create a name vector label used later in classification heatmap where the values are cell types/ clusters and names are the sample names
 
nrand = 50
sla = as.vector(stTest$cell_ontology_class)
names(sla) = as.vector(stTest$cell)
slaRand = rep("rand", nrand) 
names(slaRand) = paste("rand_", 1:nrand, sep='')
sla = append(sla, slaRand) #include in the random cells profile created

sc_hmClass(classMat = classRes_val_all,grps = sla, max=300, isBig=TRUE)
```
```{r}
plot_attr(classRes=classRes_val_all, sampTab=stTest, nrand=nrand, dLevel="cell_ontology_class", sid="cell")
```
```{r}
gpTab = compareGenePairs(query_exp = expTest, training_exp = expTrain, 
                         training_st = stTrain, classCol = "cell_ontology_class", 
                         sampleCol = "cell", RF_classifier = class_info$cnProc$classifier, 
                         numPairs = 20, trainingOnly= TRUE)

train = findAvgLabel(gpTab = gpTab, stTrain = stTrain, dLevel = "cell_ontology_class")

hm_gpa_sel(gpTab, genes = class_info$cnProc$xpairs, grps = train, maxPerGrp = 50)
```

```{r}
expPark = seurat$expDat 
```

```{r}
nqRand = 50
system.time(crParkall<-scn_predict(class_info[['cnProc']], expPark, nrand=nqRand))
```

```{r}
sgrp = as.vector(stPark$seurat_clusters)
names(sgrp) = as.vector(rownames(stPark))
grpRand =rep("rand", nqRand)
names(grpRand) = paste("rand_", 1:nqRand, sep='')
sgrp = append(sgrp, grpRand)

sc_hmClass(crParkall, sgrp, max=5000, isBig=TRUE, cCol=F, font=8)
```

```{r}
FIG <- sc_hmClass(crParkall, sgrp, max=5000, isBig=TRUE, cCol=F, font=8)
ggsave(FIG, filename="... .png", width=12, height=7, dpi=900)
```

```{r}
expPark = seurat$expDat 
```

```{r}
system.time(crParkall<-scn_predict(class_info[['cnProc']], expPark, nrand = F))
```

```{r}
#lets create a column in the stPark dataframe called SID with the rownames 
stPark$SID <- as.vector(rownames(stPark))
```

```{r}
stPark <- get_cate(classRes = crParkall, sampTab = stPark, dLevel = "seurat_clusters", 
                   sid = "SID", nrand = nqRand)
stPark
```

```{r}
table(stPark[which(stPark$seurat_clusters == "0"), "category"])
table(stPark[which(stPark$seurat_clusters == "1"), "category"])
table(stPark[which(stPark$seurat_clusters == "2"), "category"])
table(stPark[which(stPark$seurat_clusters == "3"), "category"])
table(stPark[which(stPark$seurat_clusters == "4"), "category"])
table(stPark[which(stPark$seurat_clusters == "5"), "category"])
table(stPark[which(stPark$seurat_clusters == "6"), "category"])
table(stPark[which(stPark$seurat_clusters == "7"), "category"])
table(stPark[which(stPark$seurat_clusters == "8"), "category"])
table(stPark[which(stPark$seurat_clusters == "9"), "category"])
table(stPark[which(stPark$seurat_clusters == "10"), "category"])
table(stPark[which(stPark$seurat_clusters == "11"), "category"])
table(stPark[which(stPark$seurat_clusters == "12"), "category"])
table(stPark[which(stPark$seurat_clusters == "13"), "category"])
table(stPark[which(stPark$seurat_clusters == "14"), "category"])
table(stPark[which(stPark$seurat_clusters == "15"), "category"])
table(stPark[which(stPark$seurat_clusters == "16"), "category"])
table(stPark[which(stPark$seurat_clusters == "17"), "category"])
table(stPark[which(stPark$seurat_clusters == "18"), "category"])
table(stPark[which(stPark$seurat_clusters == "19"), "category"])
table(stPark[which(stPark$seurat_clusters == "20"), "category"])
table(stPark[which(stPark$seurat_clusters == "21"), "category"])
table(stPark[which(stPark$seurat_clusters == "22"), "category"])
```

```{r}
sc_violinClass(sampTab = stPark, classRes = crParkall, sid = "SID", dLevel = "seurat_clusters", ncol = 2)
```

```{r}
FIG <- sc_violinClass(sampTab = stPark, classRes = crParkall, sid = "SID", dLevel = "seurat_clusters", ncol = 2)
ggsave(FIG, filename="... .png", width=12, height=7, dpi=900)
```
```{r}
#stKid2 = addRandToSampTab(crParkall, stPark, "seurat_clusters", "SID")
skylineClass(crParkall, "T cell", stKid2, "seurat_clusters",.25, "SID")
```
